+++ src/offers/components/OfferDetailBody.vue
@@ -35,7 +35,7 @@
         {{ offer.status }}
       </div>
       <div
-        v-if="offer.canEdit && offer.status === 'active'"
+        v-if="canEdit && offer.status === 'active'"
         class="row justify-end"
       >
         <QBtnDropdown
@@ -63,7 +63,10 @@
         </QBtnDropdown>
       </div>
       <div class="q-ma-md q-pa-md bg-white grey-border">
-        <Markdown :source="offer.description" />
+        <Markdown
+          v-measure
+          :source="offer.description"
+        />
       </div>
     </template>
   </ChatConversation>
@@ -76,7 +79,9 @@ import ChatConversation from '@/messages/components/ChatConversation'
 import Markdown from '@/utils/components/Markdown'
 import KSpinner from '@/utils/components/KSpinner'
 import { QBtn, QBtnDropdown, QCarousel, QCarouselSlide } from 'quasar'
-import { DEFAULT_STATUS } from '@/offers/datastore/offers'
+import { DEFAULT_STATUS, useCurrentOfferQuery } from '@/offers/queries'
+import { useCurrentUserIdRef } from '@/users/queries'
+import { useArchiveOfferMutation } from '@/offers/mutations'
 
 export default {
   components: {
@@ -94,6 +99,15 @@ export default {
       default: false,
     },
   },
+  setup () {
+    const { mutate: archive } = useArchiveOfferMutation()
+    const { offer } = useCurrentOfferQuery()
+    return {
+      archive,
+      offer,
+      currentUserId: useCurrentUserIdRef(),
+    }
+  },
   data () {
     return {
       selectedImageIndex: 0,
@@ -101,11 +115,13 @@ export default {
   },
   computed: {
     ...mapGetters({
-      offer: 'currentOffer/value',
       conversation: 'currentOffer/conversation',
       away: 'presence/toggle/away',
       currentUser: 'auth/user',
     }),
+    canEdit () {
+      return this.offer.user === this.currentUserId
+    },
     conversationWithReversedMessages () {
       return {
         ...this.conversation,
@@ -140,7 +156,6 @@ export default {
       toggleReaction: 'conversations/toggleReaction',
       fetchPast: 'conversations/fetchPast',
       saveConversation: 'conversations/maybeSave',
-      archive: 'offers/archive',
     }),
   },
 }