+++ src/offers/datastore/currentOffer.js
@@ -1,9 +1,18 @@
-import offers from '@/offers/api/offers'
+/**
+ *  currentOffer store module
+ *
+ *  With the move to vue-query this now only stores the id of the
+ *  currently selected offer, which is used to fetch the conversation.
+ *
+ *  The rest of the offer data is managed using vue-query.
+ *
+ *  When conversations have moved to vue-query this module can go away.
+ */
+
 import { createMetaModule, withMeta } from '@/utils/datastore/helpers'
 
 function initialState () {
   return {
-    current: null,
     id: null,
   }
 }
@@ -15,40 +24,12 @@ export default {
   },
   state: initialState(),
   getters: {
-    value: (state, getters, rootState, rootGetters) => {
-      return rootGetters['offers/enrich'](state.current)
-    },
-    id: state => state.id,
     conversation: (state, getters, rootState, rootGetters) => {
-      if (!state.current) return
-      return rootGetters['conversations/getForOffer'](state.current.id)
-    },
-    fetching (state, getters) {
       if (!state.id) return
-      const status = getters['meta/status']('fetch', state.id)
-      return status && status.pending
-    },
-    saveStatus: (state, getters) => {
-      const currentOffer = getters.value
-      return currentOffer && currentOffer.saveStatus
+      return rootGetters['conversations/getForOffer'](state.id)
     },
   },
   actions: {
-    ...withMeta({
-      async fetch ({ state, commit, dispatch }, offerId) {
-        commit('setId', offerId)
-        const offer = await offers.get(offerId)
-
-        // aborting, another offer has been loaded while we waited
-        if (state.id !== offerId) return
-        commit('set', offer)
-      },
-    }),
-    refresh ({ state, dispatch }) {
-      if (state.current) {
-        dispatch('fetch', state.current.id)
-      }
-    },
     clear ({ commit }) {
       commit('clear')
     },
@@ -57,8 +38,6 @@ export default {
         // clear right drawer
         // TODO can be removed once detail are bound to routes
         dispatch('detail/clear', null, { root: true })
-
-        await dispatch('fetch', offerId)
         dispatch('conversations/fetchForOffer', { offerId }, { root: true })
       },
     }, {
@@ -71,19 +50,8 @@ export default {
     setId (state, value) {
       state.id = value
     },
-    set (state, offer) {
-      state.current = Object.freeze(offer)
-    },
     clear (state) {
       Object.assign(state, initialState())
-    },
-    update (state, offer) {
-      if (state.id === offer.id) {
-        state.current = Object.freeze(offer)
-      }
-    },
-    delete (state, offerId) {
-      if (state.id === offerId) Object.assign(state, initialState())
     },
   },
 }