+++ src/offers/datastore/index.js
@@ -1,13 +1,10 @@
-import offers, { plugin as offersPlugin } from '@/offers/datastore/offers'
 import currentOffer, { plugin as currentOfferPlugin } from '@/offers/datastore/currentOffer'
 
 export default {
   modules: {
-    offers,
     currentOffer,
   },
   plugins: [
-    offersPlugin,
     currentOfferPlugin,
   ],
 }